meson.build: work around meson braindamage
It is required to manually call

ninja  -C build vapi/libmutter-10.vapi
ninja  -C build vapi/libmutter-10.deps

before calling

ninja -C build.

meson is victorious and we lie defeated. there is no way to trigger
the building of custom targets so they can be used by meson build.

it would work if a custom target also produced a dependency object
which could be used wherever, and which would then cause the custom
target to be built.  but that's not there now.